requirements.txt
email-validator>=2.2.0
flask-session>=0.8.0
flask>=3.1.2
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.43
werkzeug>=3.1.3
google-genai>=1.31.0